Moved all Exception classes to the Exception namespace and folder in the SDK.
