+++ OpenIIoT.Core.Tests/Platform/PlatformDirectories.cs
@@ -51,6 +51,7 @@
 using System;
 using System.Collections.Generic;
 using OpenIIoT.SDK.Platform;
+using OpenIIoT.SDK.Exceptions;
 using Xunit;
 
 namespace OpenIIoT.Core.Tests
