+++ OpenIIoT.SDK/Exceptions/ApplicationManagerExceptions.cs
@@ -1,14 +1,14 @@
 ﻿/*
       █▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀ ▀▀▀▀▀▀▀▀▀▀▀▀▀▀ ▀▀▀  ▀  ▀      ▀▀
       █
-      █     ▄████████
-      █     ███    ███
-      █     ███    ███    █████▄    █████▄  █        █   ▄██████   ▄█████      ██     █   ██████  ██▄▄▄▄
-      █     ███    ███   ██   ██   ██   ██ ██       ██  ██    ██   ██   ██ ▀███████▄ ██  ██    ██ ██▀▀▀█▄
-      █   ▀███████████   ██   ██   ██   ██ ██       ██▌ ██    ▀    ██   ██     ██  ▀ ██▌ ██    ██ ██   ██
-      █     ███    ███ ▀██████▀  ▀██████▀  ██       ██  ██    ▄  ▀████████     ██    ██  ██    ██ ██   ██
-      █     ███    ███   ██        ██      ██▌    ▄ ██  ██    ██   ██   ██     ██    ██  ██    ██ ██   ██
-      █     ███    █▀   ▄███▀     ▄███▀    ████▄▄██ █   ██████▀    ██   █▀    ▄██▀   █    ██████   █   █
+      █     ▄████████                                                                                        ▄▄▄▄███▄▄▄▄
+      █     ███    ███                                                                                     ▄██▀▀▀███▀▀▀██▄
+      █     ███    ███    █████▄    █████▄  █        █   ▄██████   ▄█████      ██     █   ██████  ██▄▄▄▄   ███   ███   ███   ▄█████  ██▄▄▄▄    ▄█████     ▄████▄     ▄█████    █████
+      █     ███    ███   ██   ██   ██   ██ ██       ██  ██    ██   ██   ██ ▀███████▄ ██  ██    ██ ██▀▀▀█▄  ███   ███   ███   ██   ██ ██▀▀▀█▄   ██   ██   ██    ▀    ██   █    ██  ██
+      █   ▀███████████   ██   ██   ██   ██ ██       ██▌ ██    ▀    ██   ██     ██  ▀ ██▌ ██    ██ ██   ██  ███   ███   ███   ██   ██ ██   ██   ██   ██  ▄██        ▄██▄▄     ▄██▄▄█▀
+      █     ███    ███ ▀██████▀  ▀██████▀  ██       ██  ██    ▄  ▀████████     ██    ██  ██    ██ ██   ██  ███   ███   ███ ▀████████ ██   ██ ▀████████ ▀▀██ ███▄  ▀▀██▀▀    ▀███████
+      █     ███    ███   ██        ██      ██▌    ▄ ██  ██    ██   ██   ██     ██    ██  ██    ██ ██   ██  ███   ███   ███   ██   ██ ██   ██   ██   ██   ██    ██   ██   █    ██  ██
+      █     ███    █▀   ▄███▀     ▄███▀    ████▄▄██ █   ██████▀    ██   █▀    ▄██▀   █    ██████   █   █    ▀█   ███   █▀    ██   █▀  █   █    ██   █▀   ██████▀    ███████   ██  ██
       █
       █      ▄████████
       █     ███    ███
@@ -22,7 +22,7 @@
  ▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄ ▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄ ▄▄  ▄▄ ▄▄   ▄▄▄▄ ▄▄     ▄▄     ▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄ ▄ ▄
  █████████████████████████████████████████████████████████████ ███████████████ ██  ██ ██   ████ ██     ██     ████████████████ █ █
       ▄
-      █  Exceptions for the Program class.
+      █  Exceptions for the ApplicationManager class.
       █
       █▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀ ▀▀▀▀▀▀▀▀▀▀▀ ▀ ▀▀▀     ▀▀               ▀
       █  The GNU Affero General Public License (GNU AGPL)
@@ -51,42 +51,42 @@
 using System;
 using System.Diagnostics.CodeAnalysis;
 
-namespace OpenIIoT.SDK
+namespace OpenIIoT.SDK.Exceptions
 {
     /// <summary>
-    ///     Represents errors that occur while parsing and applying command line arguments.
+    ///     Represents errors that occur within the ApplicationManager class.
     /// </summary>
     [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleClass", Justification = "Reviewed.")]
     [ExcludeFromCodeCoverage]
     [Serializable]
-    public class ApplicationArgumentException : ApplicationException
+    public class ApplicationManagerException : Exception
     {
         #region Public Constructors
 
         /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationArgumentException"/> class.
+        ///     Initializes a new instance of the <see cref="ApplicationManagerException"/> class.
         /// </summary>
-        public ApplicationArgumentException() : base()
+        public ApplicationManagerException() : base()
         {
         }
 
         /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationArgumentException"/> class with a specified error message.
+        ///     Initializes a new instance of the <see cref="ApplicationManagerException"/> class with a specified error message.
         /// </summary>
         /// <param name="message">The message that describes the error.</param>
-        public ApplicationArgumentException(string message) : base(message)
+        public ApplicationManagerException(string message) : base(message)
         {
         }
 
         /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationArgumentException"/> class with a specified error message
+        ///     Initializes a new instance of the <see cref="ApplicationManagerException"/> class with a specified error message
         ///     and a reference to the inner exception that is the cause of this exception.
         /// </summary>
         /// <param name="message">The error message that explains the reason for the exception.</param>
         /// <param name="innerException">
         ///     The exception that is the cause of the current exception, or a null reference if no inner exception is specified.
         /// </param>
-        public ApplicationArgumentException(string message, Exception innerException) : base(message, innerException)
+        public ApplicationManagerException(string message, Exception innerException) : base(message, innerException)
         {
         }
 
@@ -94,119 +94,39 @@ namespace OpenIIoT.SDK
     }
 
     /// <summary>
-    ///     Represents errors that occur while initializing the Application.
+    ///     Represents errors that occur while fetching Manager dependencies.
     /// </summary>
     [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleClass", Justification = "Reviewed.")]
     [ExcludeFromCodeCoverage]
     [Serializable]
-    public class ApplicationInitializationException : ApplicationException
+    public class ManagerDependencyException : ApplicationManagerException
     {
         #region Public Constructors
 
         /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationInitializationException"/> class.
+        ///     Initializes a new instance of the <see cref="ManagerDependencyException"/> class.
         /// </summary>
-        public ApplicationInitializationException() : base()
+        public ManagerDependencyException() : base()
         {
         }
 
         /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationInitializationException"/> class with a specified error message.
-        /// </summary>
-        /// <param name="message">The message that describes the error.</param>
-        public ApplicationInitializationException(string message) : base(message)
-        {
-        }
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationInitializationException"/> class with a specified error
-        ///     message and a reference to the inner exception that is the cause of this exception.
-        /// </summary>
-        /// <param name="message">The error message that explains the reason for the exception.</param>
-        /// <param name="innerException">
-        ///     The exception that is the cause of the current exception, or a null reference if no inner exception is specified.
-        /// </param>
-        public ApplicationInitializationException(string message, Exception innerException) : base(message, innerException)
-        {
-        }
-
-        #endregion Public Constructors
-    }
-
-    /// <summary>
-    ///     Represents errors that occur while performing the Application shutdown routine.
-    /// </summary>
-    [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleClass", Justification = "Reviewed.")]
-    [ExcludeFromCodeCoverage]
-    [Serializable]
-    public class ApplicationShutdownRoutineException : ApplicationException
-    {
-        #region Public Constructors
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationShutdownRoutineException"/> class.
-        /// </summary>
-        public ApplicationShutdownRoutineException() : base()
-        {
-        }
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationShutdownRoutineException"/> class with a specified error message.
+        ///     Initializes a new instance of the <see cref="ManagerDependencyException"/> class with a specified error message.
         /// </summary>
         /// <param name="message">The message that describes the error.</param>
-        public ApplicationShutdownRoutineException(string message) : base(message)
+        public ManagerDependencyException(string message) : base(message)
         {
         }
 
         /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationShutdownRoutineException"/> class with a specified error
-        ///     message and a reference to the inner exception that is the cause of this exception.
-        /// </summary>
-        /// <param name="message">The error message that explains the reason for the exception.</param>
-        /// <param name="innerException">
-        ///     The exception that is the cause of the current exception, or a null reference if no inner exception is specified.
-        /// </param>
-        public ApplicationShutdownRoutineException(string message, Exception innerException) : base(message, innerException)
-        {
-        }
-
-        #endregion Public Constructors
-    }
-
-    /// <summary>
-    ///     Represents errors that occur while starting the Application.
-    /// </summary>
-    [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleClass", Justification = "Reviewed.")]
-    [ExcludeFromCodeCoverage]
-    [Serializable]
-    public class ApplicationStartException : ApplicationException
-    {
-        #region Public Constructors
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationStartException"/> class.
-        /// </summary>
-        public ApplicationStartException() : base()
-        {
-        }
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationStartException"/> class with a specified error message.
-        /// </summary>
-        /// <param name="message">The message that describes the error.</param>
-        public ApplicationStartException(string message) : base(message)
-        {
-        }
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationStartException"/> class with a specified error message and
+        ///     Initializes a new instance of the <see cref="ManagerDependencyException"/> class with a specified error message and
         ///     a reference to the inner exception that is the cause of this exception.
         /// </summary>
         /// <param name="message">The error message that explains the reason for the exception.</param>
         /// <param name="innerException">
         ///     The exception that is the cause of the current exception, or a null reference if no inner exception is specified.
         /// </param>
-        public ApplicationStartException(string message, Exception innerException) : base(message, innerException)
+        public ManagerDependencyException(string message, Exception innerException) : base(message, innerException)
         {
         }
 
@@ -214,39 +134,39 @@ namespace OpenIIoT.SDK
     }
 
     /// <summary>
-    ///     Represents errors that occur while performing the Application startup routine.
+    ///     Represents errors that occur when the instantiation of a Manager returns an abnormal result.
     /// </summary>
     [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleClass", Justification = "Reviewed.")]
     [ExcludeFromCodeCoverage]
     [Serializable]
-    public class ApplicationStartupRoutineException : ApplicationException
+    public class ManagerInstantiationException : ApplicationManagerException
     {
         #region Public Constructors
 
         /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationStartupRoutineException"/> class.
+        ///     Initializes a new instance of the <see cref="ManagerInstantiationException"/> class.
         /// </summary>
-        public ApplicationStartupRoutineException() : base()
+        public ManagerInstantiationException() : base()
         {
         }
 
         /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationStartupRoutineException"/> class with a specified error message.
+        ///     Initializes a new instance of the <see cref="ManagerInstantiationException"/> class with a specified error message.
         /// </summary>
         /// <param name="message">The message that describes the error.</param>
-        public ApplicationStartupRoutineException(string message) : base(message)
+        public ManagerInstantiationException(string message) : base(message)
         {
         }
 
         /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationStartupRoutineException"/> class with a specified error
-        ///     message and a reference to the inner exception that is the cause of this exception.
+        ///     Initializes a new instance of the <see cref="ManagerInstantiationException"/> class with a specified error message
+        ///     and a reference to the inner exception that is the cause of this exception.
         /// </summary>
         /// <param name="message">The error message that explains the reason for the exception.</param>
         /// <param name="innerException">
         ///     The exception that is the cause of the current exception, or a null reference if no inner exception is specified.
         /// </param>
-        public ApplicationStartupRoutineException(string message, Exception innerException) : base(message, innerException)
+        public ManagerInstantiationException(string message, Exception innerException) : base(message, innerException)
         {
         }
 
@@ -254,39 +174,119 @@ namespace OpenIIoT.SDK
     }
 
     /// <summary>
-    ///     Represents errors that occur while stopping the Application.
+    ///     Represents errors that occur when a Manager instance is requested but the Manager has not yet been initialized.
     /// </summary>
     [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleClass", Justification = "Reviewed.")]
     [ExcludeFromCodeCoverage]
     [Serializable]
-    public class ApplicationStopException : ApplicationException
+    public class ManagerNotInitializedException : ApplicationManagerException
     {
         #region Public Constructors
 
         /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationStopException"/> class.
+        ///     Initializes a new instance of the <see cref="ManagerNotInitializedException"/> class.
         /// </summary>
-        public ApplicationStopException() : base()
+        public ManagerNotInitializedException() : base()
         {
         }
 
         /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationStopException"/> class with a specified error message.
+        ///     Initializes a new instance of the <see cref="ManagerNotInitializedException"/> class with a specified error message.
         /// </summary>
         /// <param name="message">The message that describes the error.</param>
-        public ApplicationStopException(string message) : base(message)
+        public ManagerNotInitializedException(string message) : base(message)
         {
         }
 
         /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationStopException"/> class with a specified error message and a
+        ///     Initializes a new instance of the <see cref="ManagerNotInitializedException"/> class with a specified error message
+        ///     and a reference to the inner exception that is the cause of this exception.
+        /// </summary>
+        /// <param name="message">The error message that explains the reason for the exception.</param>
+        /// <param name="innerException">
+        ///     The exception that is the cause of the current exception, or a null reference if no inner exception is specified.
+        /// </param>
+        public ManagerNotInitializedException(string message, Exception innerException) : base(message, innerException)
+        {
+        }
+
+        #endregion Public Constructors
+    }
+
+    /// <summary>
+    ///     Represents errors that occur during the registration of a Manager instance.
+    /// </summary>
+    [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleClass", Justification = "Reviewed.")]
+    [ExcludeFromCodeCoverage]
+    [Serializable]
+    public class ManagerRegistrationException : ApplicationManagerException
+    {
+        #region Public Constructors
+
+        /// <summary>
+        ///     Initializes a new instance of the <see cref="ManagerRegistrationException"/> class.
+        /// </summary>
+        public ManagerRegistrationException() : base()
+        {
+        }
+
+        /// <summary>
+        ///     Initializes a new instance of the <see cref="ManagerRegistrationException"/> class with a specified error message.
+        /// </summary>
+        /// <param name="message">The message that describes the error.</param>
+        public ManagerRegistrationException(string message) : base(message)
+        {
+        }
+
+        /// <summary>
+        ///     Initializes a new instance of the <see cref="ManagerRegistrationException"/> class with a specified error message
+        ///     and a reference to the inner exception that is the cause of this exception.
+        /// </summary>
+        /// <param name="message">The error message that explains the reason for the exception.</param>
+        /// <param name="innerException">
+        ///     The exception that is the cause of the current exception, or a null reference if no inner exception is specified.
+        /// </param>
+        public ManagerRegistrationException(string message, Exception innerException) : base(message, innerException)
+        {
+        }
+
+        #endregion Public Constructors
+    }
+
+    /// <summary>
+    ///     Represents errors that occur when the type list argument for the ApplicationManager constructor is malformed.
+    /// </summary>
+    [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleClass", Justification = "Reviewed.")]
+    [ExcludeFromCodeCoverage]
+    [Serializable]
+    public class ManagerTypeListException : ApplicationManagerException
+    {
+        #region Public Constructors
+
+        /// <summary>
+        ///     Initializes a new instance of the <see cref="ManagerTypeListException"/> class.
+        /// </summary>
+        public ManagerTypeListException() : base()
+        {
+        }
+
+        /// <summary>
+        ///     Initializes a new instance of the <see cref="ManagerTypeListException"/> class with a specified error message.
+        /// </summary>
+        /// <param name="message">The message that describes the error.</param>
+        public ManagerTypeListException(string message) : base(message)
+        {
+        }
+
+        /// <summary>
+        ///     Initializes a new instance of the <see cref="ManagerTypeListException"/> class with a specified error message and a
         ///     reference to the inner exception that is the cause of this exception.
         /// </summary>
         /// <param name="message">The error message that explains the reason for the exception.</param>
         /// <param name="innerException">
         ///     The exception that is the cause of the current exception, or a null reference if no inner exception is specified.
         /// </param>
-        public ApplicationStopException(string message, Exception innerException) : base(message, innerException)
+        public ManagerTypeListException(string message, Exception innerException) : base(message, innerException)
         {
         }
