+++ OpenIIoT.SDK/Exceptions/ManagerExceptions.cs
@@ -1,14 +1,14 @@
 ﻿/*
       █▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀ ▀▀▀▀▀▀▀▀▀▀▀▀▀▀ ▀▀▀  ▀  ▀      ▀▀
       █
-      █     ▄████████                                                                                        ▄▄▄▄███▄▄▄▄
-      █     ███    ███                                                                                     ▄██▀▀▀███▀▀▀██▄
-      █     ███    ███    █████▄    █████▄  █        █   ▄██████   ▄█████      ██     █   ██████  ██▄▄▄▄   ███   ███   ███   ▄█████  ██▄▄▄▄    ▄█████     ▄████▄     ▄█████    █████
-      █     ███    ███   ██   ██   ██   ██ ██       ██  ██    ██   ██   ██ ▀███████▄ ██  ██    ██ ██▀▀▀█▄  ███   ███   ███   ██   ██ ██▀▀▀█▄   ██   ██   ██    ▀    ██   █    ██  ██
-      █   ▀███████████   ██   ██   ██   ██ ██       ██▌ ██    ▀    ██   ██     ██  ▀ ██▌ ██    ██ ██   ██  ███   ███   ███   ██   ██ ██   ██   ██   ██  ▄██        ▄██▄▄     ▄██▄▄█▀
-      █     ███    ███ ▀██████▀  ▀██████▀  ██       ██  ██    ▄  ▀████████     ██    ██  ██    ██ ██   ██  ███   ███   ███ ▀████████ ██   ██ ▀████████ ▀▀██ ███▄  ▀▀██▀▀    ▀███████
-      █     ███    ███   ██        ██      ██▌    ▄ ██  ██    ██   ██   ██     ██    ██  ██    ██ ██   ██  ███   ███   ███   ██   ██ ██   ██   ██   ██   ██    ██   ██   █    ██  ██
-      █     ███    █▀   ▄███▀     ▄███▀    ████▄▄██ █   ██████▀    ██   █▀    ▄██▀   █    ██████   █   █    ▀█   ███   █▀    ██   █▀  █   █    ██   █▀   ██████▀    ███████   ██  ██
+      █      ▄▄▄▄███▄▄▄▄
+      █    ▄██▀▀▀███▀▀▀██▄
+      █    ███   ███   ███   ▄█████  ██▄▄▄▄    ▄█████     ▄████▄     ▄█████    █████
+      █    ███   ███   ███   ██   ██ ██▀▀▀█▄   ██   ██   ██    ▀    ██   █    ██  ██
+      █    ███   ███   ███   ██   ██ ██   ██   ██   ██  ▄██        ▄██▄▄     ▄██▄▄█▀
+      █    ███   ███   ███ ▀████████ ██   ██ ▀████████ ▀▀██ ███▄  ▀▀██▀▀    ▀███████
+      █    ███   ███   ███   ██   ██ ██   ██   ██   ██   ██    ██   ██   █    ██  ██
+      █     ▀█   ███   █▀    ██   █▀  █   █    ██   █▀   ██████▀    ███████   ██  ██
       █
       █      ▄████████
       █     ███    ███
@@ -22,7 +22,7 @@
  ▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄ ▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄ ▄▄  ▄▄ ▄▄   ▄▄▄▄ ▄▄     ▄▄     ▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄ ▄ ▄
  █████████████████████████████████████████████████████████████ ███████████████ ██  ██ ██   ████ ██     ██     ████████████████ █ █
       ▄
-      █  Exceptions for the ApplicationManager class.
+      █  Exceptions for the Manager class and subclasses.
       █
       █▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀ ▀▀▀▀▀▀▀▀▀▀▀ ▀ ▀▀▀     ▀▀               ▀
       █  The GNU Affero General Public License (GNU AGPL)
@@ -51,42 +51,42 @@
 using System;
 using System.Diagnostics.CodeAnalysis;
 
-namespace OpenIIoT.SDK
+namespace OpenIIoT.SDK.Exceptions
 {
     /// <summary>
-    ///     Represents errors that occur within the ApplicationManager class.
+    ///     Represents errors that occur when referenced dependencies can not be resolved.
     /// </summary>
     [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleClass", Justification = "Reviewed.")]
     [ExcludeFromCodeCoverage]
     [Serializable]
-    public class ApplicationManagerException : Exception
+    public class DependencyNotResolvedException : ManagerException
     {
         #region Public Constructors
 
         /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationManagerException"/> class.
+        ///     Initializes a new instance of the <see cref="DependencyNotResolvedException"/> class.
         /// </summary>
-        public ApplicationManagerException() : base()
+        public DependencyNotResolvedException() : base()
         {
         }
 
         /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationManagerException"/> class with a specified error message.
+        ///     Initializes a new instance of the <see cref="DependencyNotResolvedException"/> class with a specified error message.
         /// </summary>
         /// <param name="message">The message that describes the error.</param>
-        public ApplicationManagerException(string message) : base(message)
+        public DependencyNotResolvedException(string message) : base(message)
         {
         }
 
         /// <summary>
-        ///     Initializes a new instance of the <see cref="ApplicationManagerException"/> class with a specified error message
+        ///     Initializes a new instance of the <see cref="DependencyNotResolvedException"/> class with a specified error message
         ///     and a reference to the inner exception that is the cause of this exception.
         /// </summary>
         /// <param name="message">The error message that explains the reason for the exception.</param>
         /// <param name="innerException">
         ///     The exception that is the cause of the current exception, or a null reference if no inner exception is specified.
         /// </param>
-        public ApplicationManagerException(string message, Exception innerException) : base(message, innerException)
+        public DependencyNotResolvedException(string message, Exception innerException) : base(message, innerException)
         {
         }
 
@@ -94,199 +94,159 @@ namespace OpenIIoT.SDK
     }
 
     /// <summary>
-    ///     Represents errors that occur while fetching Manager dependencies.
+    ///     Represents errors that occur at the program level.
     /// </summary>
     [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleClass", Justification = "Reviewed.")]
     [ExcludeFromCodeCoverage]
     [Serializable]
-    public class ManagerDependencyException : ApplicationManagerException
+    public class ManagerException : Exception
     {
         #region Public Constructors
 
         /// <summary>
-        ///     Initializes a new instance of the <see cref="ManagerDependencyException"/> class.
+        ///     Initializes a new instance of the <see cref="ManagerException"/> class.
         /// </summary>
-        public ManagerDependencyException() : base()
+        public ManagerException() : base()
         {
         }
 
         /// <summary>
-        ///     Initializes a new instance of the <see cref="ManagerDependencyException"/> class with a specified error message.
-        /// </summary>
-        /// <param name="message">The message that describes the error.</param>
-        public ManagerDependencyException(string message) : base(message)
-        {
-        }
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="ManagerDependencyException"/> class with a specified error message and
-        ///     a reference to the inner exception that is the cause of this exception.
-        /// </summary>
-        /// <param name="message">The error message that explains the reason for the exception.</param>
-        /// <param name="innerException">
-        ///     The exception that is the cause of the current exception, or a null reference if no inner exception is specified.
-        /// </param>
-        public ManagerDependencyException(string message, Exception innerException) : base(message, innerException)
-        {
-        }
-
-        #endregion Public Constructors
-    }
-
-    /// <summary>
-    ///     Represents errors that occur when the instantiation of a Manager returns an abnormal result.
-    /// </summary>
-    [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleClass", Justification = "Reviewed.")]
-    [ExcludeFromCodeCoverage]
-    [Serializable]
-    public class ManagerInstantiationException : ApplicationManagerException
-    {
-        #region Public Constructors
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="ManagerInstantiationException"/> class.
-        /// </summary>
-        public ManagerInstantiationException() : base()
-        {
-        }
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="ManagerInstantiationException"/> class with a specified error message.
+        ///     Initializes a new instance of the <see cref="ManagerException"/> class with a specified error message.
         /// </summary>
         /// <param name="message">The message that describes the error.</param>
-        public ManagerInstantiationException(string message) : base(message)
+        public ManagerException(string message) : base(message)
         {
         }
 
         /// <summary>
-        ///     Initializes a new instance of the <see cref="ManagerInstantiationException"/> class with a specified error message
-        ///     and a reference to the inner exception that is the cause of this exception.
-        /// </summary>
-        /// <param name="message">The error message that explains the reason for the exception.</param>
-        /// <param name="innerException">
-        ///     The exception that is the cause of the current exception, or a null reference if no inner exception is specified.
-        /// </param>
-        public ManagerInstantiationException(string message, Exception innerException) : base(message, innerException)
-        {
-        }
-
-        #endregion Public Constructors
-    }
-
-    /// <summary>
-    ///     Represents errors that occur when a Manager instance is requested but the Manager has not yet been initialized.
-    /// </summary>
-    [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleClass", Justification = "Reviewed.")]
-    [ExcludeFromCodeCoverage]
-    [Serializable]
-    public class ManagerNotInitializedException : ApplicationManagerException
-    {
-        #region Public Constructors
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="ManagerNotInitializedException"/> class.
-        /// </summary>
-        public ManagerNotInitializedException() : base()
-        {
-        }
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="ManagerNotInitializedException"/> class with a specified error message.
-        /// </summary>
-        /// <param name="message">The message that describes the error.</param>
-        public ManagerNotInitializedException(string message) : base(message)
-        {
-        }
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="ManagerNotInitializedException"/> class with a specified error message
-        ///     and a reference to the inner exception that is the cause of this exception.
-        /// </summary>
-        /// <param name="message">The error message that explains the reason for the exception.</param>
-        /// <param name="innerException">
-        ///     The exception that is the cause of the current exception, or a null reference if no inner exception is specified.
-        /// </param>
-        public ManagerNotInitializedException(string message, Exception innerException) : base(message, innerException)
-        {
-        }
-
-        #endregion Public Constructors
-    }
-
-    /// <summary>
-    ///     Represents errors that occur during the registration of a Manager instance.
-    /// </summary>
-    [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleClass", Justification = "Reviewed.")]
-    [ExcludeFromCodeCoverage]
-    [Serializable]
-    public class ManagerRegistrationException : ApplicationManagerException
-    {
-        #region Public Constructors
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="ManagerRegistrationException"/> class.
-        /// </summary>
-        public ManagerRegistrationException() : base()
-        {
-        }
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="ManagerRegistrationException"/> class with a specified error message.
-        /// </summary>
-        /// <param name="message">The message that describes the error.</param>
-        public ManagerRegistrationException(string message) : base(message)
-        {
-        }
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="ManagerRegistrationException"/> class with a specified error message
-        ///     and a reference to the inner exception that is the cause of this exception.
-        /// </summary>
-        /// <param name="message">The error message that explains the reason for the exception.</param>
-        /// <param name="innerException">
-        ///     The exception that is the cause of the current exception, or a null reference if no inner exception is specified.
-        /// </param>
-        public ManagerRegistrationException(string message, Exception innerException) : base(message, innerException)
-        {
-        }
-
-        #endregion Public Constructors
-    }
-
-    /// <summary>
-    ///     Represents errors that occur when the type list argument for the ApplicationManager constructor is malformed.
-    /// </summary>
-    [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleClass", Justification = "Reviewed.")]
-    [ExcludeFromCodeCoverage]
-    [Serializable]
-    public class ManagerTypeListException : ApplicationManagerException
-    {
-        #region Public Constructors
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="ManagerTypeListException"/> class.
-        /// </summary>
-        public ManagerTypeListException() : base()
-        {
-        }
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="ManagerTypeListException"/> class with a specified error message.
-        /// </summary>
-        /// <param name="message">The message that describes the error.</param>
-        public ManagerTypeListException(string message) : base(message)
-        {
-        }
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="ManagerTypeListException"/> class with a specified error message and a
+        ///     Initializes a new instance of the <see cref="ManagerException"/> class with a specified error message and a
         ///     reference to the inner exception that is the cause of this exception.
         /// </summary>
         /// <param name="message">The error message that explains the reason for the exception.</param>
         /// <param name="innerException">
         ///     The exception that is the cause of the current exception, or a null reference if no inner exception is specified.
         /// </param>
-        public ManagerTypeListException(string message, Exception innerException) : base(message, innerException)
+        public ManagerException(string message, Exception innerException) : base(message, innerException)
+        {
+        }
+
+        #endregion Public Constructors
+    }
+
+    /// <summary>
+    ///     Represents errors that occur when the setup of a Manager returns an abnormal result.
+    /// </summary>
+    [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleClass", Justification = "Reviewed.")]
+    [ExcludeFromCodeCoverage]
+    [Serializable]
+    public class ManagerSetupException : ManagerException
+    {
+        #region Public Constructors
+
+        /// <summary>
+        ///     Initializes a new instance of the <see cref="ManagerSetupException"/> class.
+        /// </summary>
+        public ManagerSetupException() : base()
+        {
+        }
+
+        /// <summary>
+        ///     Initializes a new instance of the <see cref="ManagerSetupException"/> class with a specified error message.
+        /// </summary>
+        /// <param name="message">The message that describes the error.</param>
+        public ManagerSetupException(string message) : base(message)
+        {
+        }
+
+        /// <summary>
+        ///     Initializes a new instance of the <see cref="ManagerSetupException"/> class with a specified error message and a
+        ///     reference to the inner exception that is the cause of this exception.
+        /// </summary>
+        /// <param name="message">The error message that explains the reason for the exception.</param>
+        /// <param name="innerException">
+        ///     The exception that is the cause of the current exception, or a null reference if no inner exception is specified.
+        /// </param>
+        public ManagerSetupException(string message, Exception innerException) : base(message, innerException)
+        {
+        }
+
+        #endregion Public Constructors
+    }
+
+    /// <summary>
+    ///     Represents errors that occur when a Manager fails to start.
+    /// </summary>
+    [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleClass", Justification = "Reviewed.")]
+    [ExcludeFromCodeCoverage]
+    [Serializable]
+    public class ManagerStartException : ManagerException
+    {
+        #region Public Constructors
+
+        /// <summary>
+        ///     Initializes a new instance of the <see cref="ManagerStartException"/> class.
+        /// </summary>
+        public ManagerStartException() : base()
+        {
+        }
+
+        /// <summary>
+        ///     Initializes a new instance of the <see cref="ManagerStartException"/> class with a specified error message.
+        /// </summary>
+        /// <param name="message">The message that describes the error.</param>
+        public ManagerStartException(string message) : base(message)
+        {
+        }
+
+        /// <summary>
+        ///     Initializes a new instance of the <see cref="ManagerStartException"/> class with a specified error message and a
+        ///     reference to the inner exception that is the cause of this exception.
+        /// </summary>
+        /// <param name="message">The error message that explains the reason for the exception.</param>
+        /// <param name="innerException">
+        ///     The exception that is the cause of the current exception, or a null reference if no inner exception is specified.
+        /// </param>
+        public ManagerStartException(string message, Exception innerException) : base(message, innerException)
+        {
+        }
+
+        #endregion Public Constructors
+    }
+
+    /// <summary>
+    ///     Represents errors that occur when a Manager fails to stop.
+    /// </summary>
+    [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleClass", Justification = "Reviewed.")]
+    [ExcludeFromCodeCoverage]
+    [Serializable]
+    public class ManagerStopException : ManagerException
+    {
+        #region Public Constructors
+
+        /// <summary>
+        ///     Initializes a new instance of the <see cref="ManagerStopException"/> class.
+        /// </summary>
+        public ManagerStopException() : base()
+        {
+        }
+
+        /// <summary>
+        ///     Initializes a new instance of the <see cref="ManagerStopException"/> class with a specified error message.
+        /// </summary>
+        /// <param name="message">The message that describes the error.</param>
+        public ManagerStopException(string message) : base(message)
+        {
+        }
+
+        /// <summary>
+        ///     Initializes a new instance of the <see cref="ManagerStopException"/> class with a specified error message and a
+        ///     reference to the inner exception that is the cause of this exception.
+        /// </summary>
+        /// <param name="message">The error message that explains the reason for the exception.</param>
+        /// <param name="innerException">
+        ///     The exception that is the cause of the current exception, or a null reference if no inner exception is specified.
+        /// </param>
+        public ManagerStopException(string message, Exception innerException) : base(message, innerException)
         {
         }
