+++ OpenIIoT.SDK/OpenIIoT.SDK.csproj
@@ -86,11 +86,11 @@
     <Compile Include="Common\Provider\EventProvider\EventAttribute.cs" />
     <Compile Include="Common\Discovery\DiscoverableAttribute.cs" />
     <Compile Include="Event\IEventManager.cs" />
-    <Compile Include="ApplicationExceptions.cs" />
-    <Compile Include="ApplicationManagerExceptions.cs" />
+    <Compile Include="Exceptions\ApplicationExceptions.cs" />
+    <Compile Include="Exceptions\ApplicationManagerExceptions.cs" />
     <Compile Include="Common\Provider\EventProvider\IEventProvider.cs" />
-    <Compile Include="Platform\PlatformExceptions.cs" />
-    <Compile Include="Common\Manager\ManagerExceptions.cs" />
+    <Compile Include="Exceptions\PlatformExceptions.cs" />
+    <Compile Include="Exceptions\ManagerExceptions.cs" />
     <Compile Include="IApplicationManager.cs" />
     <Compile Include="Model\IModelManager.cs" />
     <Compile Include="Platform\IPlatform.cs" />
